chore: make filter list server component
