+++ src/app/datasets/FilterList/index.tsx
@@ -1,30 +1,19 @@
 // SPDX-FileCopyrightText: 2024 PNED G.I.E.
 //
 // SPDX-License-Identifier: Apache-2.0
-"use client";
 
 import {
   FilterItemProps,
   convertDataToFilterItemProps,
 } from "@/utils/convertDataToFilterItemProps";
-import Button from "@/components/Button";
 import FilterItem from "./FilterItem";
 import { Facet } from "@/services/discovery/types/facets.type";
-import { useEffect, useState } from "react";
-import { useSearchParams } from "next/navigation";
+import ClearFilterButton from "./ClearFilterButton";
+import { GET } from "@/app/api/facets/route";
 
-export default function FilterList() {
-  const queryParams = useSearchParams();
-  const [searchFacets, setSearchFacets] = useState<Facet[]>([]);
-
-  useEffect(() => {
-    async function fetchSearchFacets() {
-      const response = await fetch("/api/facets");
-      const searchFacets = (await response.json()) as Facet[];
-      setSearchFacets(searchFacets);
-    }
-    fetchSearchFacets();
-  }, []);
+export default async function FilterList() {
+  const response = await GET();
+  const searchFacets = (await response.json()) as Facet[];
 
   const facetGroups = Array.from(
     new Set(searchFacets.map((facet) => facet.facetGroup))
@@ -33,27 +22,6 @@ export default function FilterList() {
   const filterItemProps: FilterItemProps[] = convertDataToFilterItemProps(
     searchFacets
   ).sort((f1, f2) => f2.groupKey.localeCompare(f1.groupKey));
-
-  function isAnyGroupFilterApplied() {
-    if (!queryParams) return false;
-    return Array.from(queryParams.keys()).some(
-      (key) =>
-        key !== "page" &&
-        key !== "q" &&
-        facetGroups.some((group) => key.includes(group))
-    );
-  }
-
-  function getQueryStringWithoutGroupFilter() {
-    const filteredParamsQuery = Array.from(queryParams.keys())
-      .filter(
-        (x) => facetGroups.every((group) => !x.includes(group)) && x !== "page"
-      )
-      .map((x) => `&${x}=${queryParams.get(x)}`)
-      .join("");
-
-    return filteredParamsQuery;
-  }
 
   return (
     <div className="flex flex-col gap-y-6">
@@ -67,15 +35,7 @@ export default function FilterList() {
           />
         </li>
       ))}
-      {isAnyGroupFilterApplied() && (
-        <div className="mt-4 flex justify-end">
-          <Button
-            href={`/datasets?page=1${getQueryStringWithoutGroupFilter()}`}
-            text="Clear Filters"
-            type="warning"
-          />
-        </div>
-      )}
+      <ClearFilterButton facetGroups={facetGroups} />
     </div>
   );
 }
